:recycle: refactored Logging into its own Impl ready for Loki integration
