+++ src/main/java/net/boomerangplatform/controller/LogController.java
@@ -1,6 +1,7 @@
 package net.boomerangplatform.controller;
 
 import javax.servlet.http.HttpServletResponse;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -9,24 +10,24 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.mvc.method.annotation.StreamingResponseBody;
-import net.boomerangplatform.model.Response;
-import net.boomerangplatform.service.AbstractControllerService;
+
+import net.boomerangplatform.service.LogService;
 
 @RestController
 @RequestMapping("/controller/log")
 public class LogController {
 
   @Autowired
-  private AbstractControllerService controllerService;
+  private LogService logService;
 
-  @GetMapping(value = "/get")
-  public Response getLogForTask(
-      @RequestParam(value = "workflowId", required = true) String workflowId,
-      @RequestParam(value = "workflowActivityId", required = true) String workflowActivityId,
-      @RequestParam(value = "taskActivityId", required = false) String taskActivityId,
-      @RequestParam(value = "taskId", required = true) String taskId) {
-    return controllerService.getLogForTask(workflowId, workflowActivityId, taskId,taskActivityId);
-  }
+//  @GetMapping(value = "/get")
+//  public Response getLogForTask(
+//      @RequestParam(value = "workflowId", required = true) String workflowId,
+//      @RequestParam(value = "workflowActivityId", required = true) String workflowActivityId,
+//      @RequestParam(value = "taskActivityId", required = false) String taskActivityId,
+//      @RequestParam(value = "taskId", required = true) String taskId) {
+//    return logService.getLogForTask(workflowId, workflowActivityId, taskId,taskActivityId);
+//  }
 
   @GetMapping(value = "/stream")
   public ResponseEntity<StreamingResponseBody> streamLogForTask(HttpServletResponse response,
@@ -35,7 +36,7 @@ public class LogController {
       @RequestParam(value = "taskActivityId", required = false) String taskActivityId,
       @RequestParam(value = "taskId", required = true) String taskId) {
     return new ResponseEntity<>(
-        controllerService.streamLogForTask(response, workflowId, workflowActivityId, taskId, taskActivityId),
+        logService.streamLogForTask(response, workflowId, workflowActivityId, taskId, taskActivityId),
         HttpStatus.OK);
   }
 }
